HUD, sounds, camera looping, more shop, bug fixes
I added sounds from a free UE4 sound pack I found in the Epic store. I also added looping so if the player leaves the frame they are spit back in on the other side.

I fixed the direct enemy not moving on spawn, as well as implementing all the shop upgrades with the actual multipliers and economy. There is a new upgrade called full auto as well where you can hold shoot, and it's very expensive.

I scaled up the planet to 10x instead of 5x as it seemed a bit too small before. I also moved the camera farther back so the game feels more like space. Coins got a rework as well, and I changed most of the enemies to have emissive materials for visibility.

Cleaned up the main HUD and added money, and made the Peace Period text larger. There is also a hacky fix so that Player 1 doesn't have to click any buttons when loading into the game, will elaborate in next meeting.

If the ship dies, it can now no longer move or shoot. We should make it float back to the shop position so that if you survive with just the planet until the shop comes back you can collide and purchase ship health.

Assuming there are no bugs, the game should now be fully playable.